fixed oauth users not getting settings created on registration
